feat: to Sign up activity
