+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools" >
 
     <application
         android:allowBackup="true"
@@ -11,10 +11,13 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.GlucoScan"
-        tools:targetApi="31">
+        tools:targetApi="31" >
         <activity
-            android:name=".view.SignInActivity"
-            android:exported="true">
+            android:name=".view.login.SignUpActivity"
+            android:exported="false" />
+        <activity
+            android:name=".view.login.SignInActivity"
+            android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -23,7 +26,7 @@
         </activity>
         <activity
             android:name=".view.MainActivity"
-            android:exported="true">
+            android:exported="true" >
         </activity>
     </application>
