+++ app/src/main/java/com/dicoding/glucoscan/view/login/SignUpActivity.kt
@@ -1,0 +1,23 @@
+package com.dicoding.glucoscan.view.login
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import com.dicoding.glucoscan.R
+import com.dicoding.glucoscan.databinding.ActivitySignUpBinding
+
+class SignUpActivity : AppCompatActivity() {
+    private lateinit var binding: ActivitySignUpBinding
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivitySignUpBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        binding.btnSignin.setOnClickListener {
+            finish()
+        }
+    }
+}
