+++ app/src/main/res/layout/activity_sign_up.xml
@@ -6,8 +6,8 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="16dp"
-    tools:context=".view.login.SignInActivity">
-    
+    tools:context=".view.login.SignUpActivity">
+
     <ImageView
         android:id="@+id/ivLogo"
         android:layout_width="64dp"
@@ -26,7 +26,7 @@
         android:textSize="24sp"
         android:textStyle="bold"
         android:layout_marginVertical="16dp"
-        android:text="@string/sign_in"
+        android:text="@string/sign_up"
         app:layout_constraintTop_toBottomOf="@id/ivLogo"
         app:layout_constraintStart_toStartOf="parent"/>
 
@@ -74,7 +74,7 @@
             android:layout_width="match_parent"
             android:layout_height="44dp"
             android:padding="8dp"
-            android:textSize="16sp"/>
+            android:textSize="16sp" />
     </com.google.android.material.textfield.TextInputLayout>
 
     <TextView
@@ -98,17 +98,32 @@
             android:layout_width="match_parent"
             android:layout_height="44dp"
             android:padding="8dp"
-            android:textSize="16sp"/>
+            android:textSize="16sp" />
     </com.google.android.material.textfield.TextInputLayout>
-
     <TextView
-        android:id="@+id/tvForgotPassword"
+        android:id="@+id/titlePasswordConfirmation"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/forgot_password"
-        android:textColor="@color/blue"
-        app:layout_constraintTop_toBottomOf="@id/tilPassword"
-        app:layout_constraintEnd_toEndOf="parent"/>
+        android:layout_marginTop="24dp"
+        app:layout_constraintStart_toStartOf="parent"
+        android:text="@string/password_confirmation"
+        app:layout_constraintTop_toBottomOf="@id/tilPassword"/>
+    <com.google.android.material.textfield.TextInputLayout
+        android:id="@+id/tilPasswordConfirmation"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        app:layout_constraintTop_toBottomOf="@id/titlePasswordConfirmation"
+        android:layout_marginTop="4dp"
+        app:hintAnimationEnabled="false"
+        app:counterEnabled="false">
+        <com.google.android.material.textfield.TextInputEditText
+            android:id="@+id/tietPasswordConfirmation"
+            android:layout_width="match_parent"
+            android:layout_height="44dp"
+            android:padding="8dp"
+            android:textSize="16sp" />
+    </com.google.android.material.textfield.TextInputLayout>
+
 
     <Button
         android:id="@+id/btnEmail"
@@ -130,16 +145,15 @@
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@string/don_t_have_an_account" />
+            android:text="@string/already_have_an_account" />
 
         <TextView
-            android:id="@+id/btnSignup"
+            android:id="@+id/btnSignin"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="4dp"
-            android:text="@string/sign_up"
+            android:text="@string/sign_in"
             android:textColor="@color/blue" />
     </LinearLayout>
-
 
 </androidx.constraintlayout.widget.ConstraintLayout>
