+++ app/src/main/res/values/strings.xml
@@ -2,5 +2,14 @@
     <string name="app_name">GlucoScan</string>
     <string name="continue_with_google">Continue with Google</string>
     <string name="sign_in">Sign in</string>
+    <string name="sign_up">Sign up</string>
+    <string name="forgot_password">Forgot password?</string>
+    <string name="continue_with_email">Continue with email</string>
+    <string name="don_t_have_an_account">Don\'t have an account? </string>"
+    <string name="already_have_an_account">Already have an account? </string>
+    <string name="email">Email</string>
+    <string name="password">Password</string>
     <string name="or">Or</string>
+    <string name="logo">Logo</string>
+    <string name="password_confirmation">Password Confirmation</string>
 </resources>
